requirements.txt
Flask
requests
gunicorn
discord.py
flask-socketio
psycopg2-binary